Fix some bugs, write some tests
