+++ tests/test_jazzmin_menus.py
@@ -1,9 +1,6 @@
 import pytest
-from bs4 import BeautifulSoup
-from django.contrib.admin.models import LogEntry, CHANGE
 from django.urls import reverse
 
-from jazzmin.templatetags import jazzmin
 from tests.utils import parse_sidemenu, user_with_permissions, parse_topmenu, parse_usermenu
 
 
@@ -48,20 +45,6 @@ def test_side_menu(admin_client, settings):
         "Administration": ["/admin/admin/logentry/"],
         "Authentication and Authorization": ["/admin/auth/group/"],
     }
-
-
-@pytest.mark.django_db
-def test_update_site_logo(admin_client, settings):
-    """
-    We can add a site logo, and it renders out
-    """
-    url = reverse("admin:index")
-
-    settings.JAZZMIN_SETTINGS["site_logo"] = "polls/img/logo.png"
-    response = admin_client.get(url)
-    soup = BeautifulSoup(response.content, "html.parser")
-
-    assert soup.find("a", class_="brand-link").find("img")["src"] == "/static/polls/img/logo.png"
 
 
 @pytest.mark.django_db
@@ -113,30 +96,12 @@ def test_top_menu(admin_client, settings):
             "name": "Polls",
             "link": "#",
             "children": [
-                {
-                    "name": "Polls",
-                    "link": reverse("admin:polls_poll_changelist")
-                },
-                {
-                    "name": "Choices",
-                    "link": reverse("admin:polls_choice_changelist")
-                },
-                {
-                    "name": "Votes",
-                    "link": reverse("admin:polls_vote_changelist")
-                },
-                {
-                    "name": "Cheeses",
-                    "link": reverse("admin:polls_cheese_changelist")
-                },
-                {
-                    "name": "Campaigns",
-                    "link": reverse("admin:polls_campaign_changelist")
-                },
-                {
-                    "name": "Allfields",
-                    "link": reverse("admin:polls_allfields_changelist")
-                },
+                {"name": "Polls", "link": reverse("admin:polls_poll_changelist")},
+                {"name": "Choices", "link": reverse("admin:polls_choice_changelist")},
+                {"name": "Votes", "link": reverse("admin:polls_vote_changelist")},
+                {"name": "Cheeses", "link": reverse("admin:polls_cheese_changelist")},
+                {"name": "Campaigns", "link": reverse("admin:polls_campaign_changelist")},
+                {"name": "Allfields", "link": reverse("admin:polls_allfields_changelist")},
             ],
         },
     ]
@@ -166,73 +131,3 @@ def test_user_menu(admin_user, client, settings):
         {"link": "/admin/auth/user/", "name": "Users"},
         {"link": "/admin/auth/user/{}/change/".format(admin_user.pk), "name": "See Profile"},
     ]
-
-
-@pytest.mark.django_db
-def test_app_is_installed(settings):
-    """
-    Returns True if an app is under INSTALLED_APPS, False otherwise
-    """
-    app = "test_app"
-
-    assert jazzmin.app_is_installed(app) is False
-
-    settings.INSTALLED_APPS.append(app)
-
-    assert jazzmin.app_is_installed(app) is True
-
-
-@pytest.mark.django_db
-def test_action_message_to_list(admin_user):
-    """
-
-    """
-    message = (
-        '[{"changed": {"fields": ["Owner", "Text", "Pub date", "Active"]}}, '
-        '{"added": {"name": "choice", "object": "More random choices"}}, '
-        '{"deleted": {"name": "choice", "object": "Person serious choose tea"}}]'
-    )
-    log_entry = LogEntry.objects.create(user=admin_user, action_flag=CHANGE, change_message=message)
-    assert jazzmin.action_message_to_list(log_entry) == [
-        "Changed Owner, Text, Pub date and Active.",
-        "Added choice “More random choices”.",
-        "Deleted choice “Person serious choose tea”.",
-    ]
-
-
-@pytest.mark.django_db
-def test_get_action_icon():
-    """
-    Returns icon depending on action being added, changed or deleted
-    """
-    add_message = "Added cheese option"
-    change_message = "Changed cheese option"
-    delete_message = "Deleted cheese option"
-
-    assert jazzmin.get_action_icon(add_message) == "plus-circle"
-    assert jazzmin.get_action_icon(change_message) == "edit"
-    assert jazzmin.get_action_icon(delete_message) == "trash"
-
-
-@pytest.mark.django_db
-def test_get_action_color():
-    """
-    Returns color depending on action being added, changed or deleted
-    """
-    add_message = "Added cheese option"
-    change_message = "Changed cheese option"
-    delete_message = "Deleted cheese option"
-
-    assert jazzmin.get_action_color(add_message) == "success"
-    assert jazzmin.get_action_color(change_message) == "blue"
-    assert jazzmin.get_action_color(delete_message) == "danger"
-
-
-@pytest.mark.django_db
-def test_style_bold_first_word():
-    """
-    Adds <strong> HTML element wrapping first word given a sentence
-    """
-    message = "The bomb has been planted"
-
-    assert jazzmin.style_bold_first_word(message) == "<strong>The</strong> bomb has been planted"
